requirements.txt
tensorflow-cpu==2.2.0
gunicorn==19.9.0
Flask==1.1.2
Jinja2==2.10.1
tifffile==2020.6.3
numpy>=1.19.0
pandas>=1.1.0
opencv-python>=4.3.0
matplotlib>=3.3.3
Pillow>=8.0.1
scikit-learn>=0.19.1
scikit-image>=0.17.2
opencv-python-headless==4.2.0.32
